Refactoring reusable components
