+++ asset-manifest.json
@@ -1,9 +1,9 @@
 {
   "files": {
-    "static/js/0.36973d74.chunk.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/0.36973d74.chunk.js",
-    "static/js/0.36973d74.chunk.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/0.36973d74.chunk.js.map",
-    "main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.26cb4874.chunk.js",
-    "main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.26cb4874.chunk.js.map",
+    "static/js/1.9b3437b3.chunk.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/1.9b3437b3.chunk.js",
+    "static/js/1.9b3437b3.chunk.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/1.9b3437b3.chunk.js.map",
+    "main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.dfe3c87e.chunk.js",
+    "main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/main.dfe3c87e.chunk.js.map",
     "runtime~main.js": "/react-wp/wp-content/themes/inu-v2-b/static/js/runtime~main.ad097248.js",
     "runtime~main.js.map": "/react-wp/wp-content/themes/inu-v2-b/static/js/runtime~main.ad097248.js.map",
     "index.php": "/react-wp/wp-content/themes/inu-v2-b/index.php"
