+++ react-src/src/components/ImageResponsive/index.js
@@ -1,0 +1,2 @@
+export { default } from "./ImageResponsive";
+export { propTypes, defaultProps } from "./ImageResponsive";
