+++ react-src/src/components/PlaceholderImage/PlaceholderImage.md
@@ -1,0 +1,1 @@
+## PlaceholderImage
