+++ react-src/src/components/PlaceholderImage/index.js
@@ -1,0 +1,6 @@
+export { default } from "./PlaceholderImage";
+export {
+  PlaceholderImagePropTypes,
+  PlaceholderImageDefaultProps,
+  createPlaceholderImageUrl
+} from "./PlaceholderImage";
